# fix #16434, superlinear compiler run time on large functions